+++ app/Http/Controllers/Accounting/MoneyTransactionsController.php
@@ -83,6 +83,7 @@ class MoneyTransactionsController extends Controller
         return view('accounting.transactions.create', [
             'beneficiaries' => self::getBeneficiaries(),
             'projects' => self::getProjects(),
+            'descriptions' => self::getDescriptions(),
             'newReceiptNo' => optional(MoneyTransaction
                 ::select(DB::raw('MAX(receipt_no) as val'))
                 ->first())
@@ -108,6 +109,17 @@ class MoneyTransactionsController extends Controller
             ->orderBy('project')
             ->get()
             ->pluck('project')
+            ->unique()
+            ->toArray();
+    }
+
+    private static function getDescriptions() {
+        return MoneyTransaction
+            ::select('description')
+            ->groupBy('description')
+            ->orderBy('description')
+            ->get()
+            ->pluck('description')
             ->unique()
             ->toArray();
     }
@@ -175,6 +187,7 @@ class MoneyTransactionsController extends Controller
             'transaction' => $transaction,
             'beneficiaries' => self::getBeneficiaries(),
             'projects' => self::getProjects(),
+            'descriptions' => self::getDescriptions(),
         ]);
     }